SavingsAccount implement 작성

+++ src/depositAccount.java
@@ -1,5 +1,5 @@
 public class DepositAccount extends BankAccount {
-    public DepositAccount(boolean newWDS, String newAccountNum, String newAccountHolder, int newAccountBalance) {
+    public DepositAccount(boolean newWDS, String newAccountNum, String newAccountHolder, double newAccountBalance) {
         super(newWDS, newAccountNum, newAccountHolder, newAccountBalance);
     }
 }
